pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imgsearch"
version = "0.1.0"
description = "Semantic image search CLI for macOS, accelerated by MLX"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "toxu" }]
keywords = ["mlx", "clip", "siglip", "image-search", "semantic-search", "macos"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "mlx>=0.20",
    "mlx-embeddings>=0.0.5",
    "faiss-cpu>=1.8.0",
    "Pillow>=10.3",
    "pillow-heif>=0.18",
    "typer>=0.12",
    "rich>=13.7",
    "numpy>=1.26,<3",
    "huggingface-hub>=0.25",
]

[project.scripts]
imgsearch = "imgsearch.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[project.urls]
Homepage = "https://github.com/toxu/imgsearch"
Issues = "https://github.com/toxu/imgsearch/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/imgsearch"]

[tool.hatch.build.targets.sdist]
include = ["src/imgsearch", "README.md", "LICENSE", "pyproject.toml"]

[tool.uv]
# Keep MLX wheels pinned to macOS arm64 builds
environments = ["sys_platform == 'darwin' and platform_machine == 'arm64'"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
